finished
Complete the given program (part2.cpp) to calculate the hitbox size of PC game characters. A
hitbox is an invisible rectangle surrounding a character and is used to detect collisions. Each
character has a length and width (in centimeters) and the product of this length and width is the
hitbox size (in square centimeters).